Inline pattern_parser mod into version_extractor
