+++ src/lib.rs
@@ -1,7 +1,6 @@
 pub mod docker_compose;
 pub mod dockerfile;
 pub mod image;
-pub mod pattern_parser;
 pub mod report;
 pub mod tag_fetcher;
 pub mod version_extractor;
